Moved static_ri dependent code from dace to static_ri. Refined static_ri/sampling.rs
