+++ dace/src/arybase.rs
@@ -1,13 +1,10 @@
-use crate::ast::{LoopBound, LoopStmt, Node, Stmt};
+use std::collections::HashMap;
+use std::rc::Rc;
+
+use crate::ast::{Node, Stmt};
 use crate::iter::Walk;
 
-use rand::prelude::*;
-use std::collections::{BTreeSet, HashMap};
-use std::intrinsics::ceilf32;
-use std::ops::Range;
-use std::rc::Rc;
-use tracing::debug;
-
+//Note: may need to change the whole program so that it does not use iterators but use other ways so that we can mute the node in rc.
 pub fn set_arybase(aloop: &mut Rc<Node>) -> (HashMap<String, usize>, usize) {
     let init = (HashMap::<String, usize>::new(), 0);
     Walk::new(aloop)
@@ -26,93 +23,6 @@ pub fn set_arybase(aloop: &mut Rc<Node>) -> (HashMap<String, usize>, usize) {
             *my_base = Some(*ary_base);
             (tbl, cur_base)
         })
-}
-
-pub fn sample_collect<'a>(
-    code_tree: &'a Node,
-    wrapping_loops: &mut Vec<&'a LoopStmt>,
-    ans: &mut HashMap<usize, Vec<(&'a str, Range<usize>)>>,
-    // access_name, (loop_name + sample_times)* + -
-    ref_counter: &mut usize,
-) {
-    // let init = (HashMap::<String, usize>::new(), 0);
-    match &code_tree.stmt {
-        Stmt::Loop(stmt) => {
-            wrapping_loops.push(stmt);
-            for i in stmt.body.iter() {
-                sample_collect(i, wrapping_loops, ans, ref_counter);
-            }
-            wrapping_loops.pop();
-        }
-        Stmt::Ref(_) => {
-            let accesses: Vec<_> = wrapping_loops
-                .iter()
-                .filter_map(|x| {
-                    let LoopBound::Fixed(lb) = x.lb else {
-                        return None;
-                    };
-                    let LoopBound::Fixed(ub) = x.ub else {
-                        return None;
-                    };
-                    Some((x.iv.as_str(), lb as usize..ub as usize))
-                })
-                .collect();
-            // let ary_name = x.name;
-            // we could use this to provide more information...
-            ans.insert(*ref_counter, accesses);
-            *ref_counter += 1;
-        }
-        Stmt::Block(x) => {
-            for i in x.iter() {
-                sample_collect(i, wrapping_loops, ans, ref_counter);
-            }
-        }
-        Stmt::Branch(_) => unimplemented!("branch is not supported yet"),
-    }
-}
-
-pub fn sample_gen(
-    collected: &mut HashMap<usize, Vec<(&str, Range<usize>)>>,
-    sampling_rate: f32,
-) -> HashMap<usize, BTreeSet<Vec<usize>>> {
-    let mut intermidiate = HashMap::<usize, f32>::new();
-    collected.iter_mut().for_each(|(ref_id, accesses)| {
-        let mut sample_times = 1.;
-        accesses.iter_mut().for_each(|(_, range)| {
-            let (lb, ub) = (range.start, range.end);
-            sample_times *= (ub - lb) as f32 * sampling_rate;
-        });
-        intermidiate.insert(*ref_id, sample_times);
-    });
-    debug!("collected: {:#?}", collected);
-    debug!("intermidiate: {:#?}", intermidiate);
-
-    let mut ans = HashMap::<usize, BTreeSet<Vec<usize>>>::new();
-    collected.iter_mut().for_each(|(ref_id, accesses)| {
-        while ans.get(ref_id).unwrap_or(&BTreeSet::new()).len()
-            < unsafe { ceilf32(*intermidiate.get(ref_id).unwrap()) as usize }
-        // FIXME: no sure if f32 is enough here
-        {
-            let mut sample_name: Vec<usize> = Vec::new();
-            accesses.iter_mut().for_each(|(_, range)| {
-                let (lb, ub) = (range.start, range.end);
-                let mut rng = rand::thread_rng();
-                let dist = rand::distributions::Uniform::new(0, ub - lb);
-                let rand_num = dist.sample(&mut rng);
-                sample_name.push(rand_num);
-            });
-            ans.entry(*ref_id)
-                .or_insert(BTreeSet::new())
-                .insert(sample_name);
-            // kinda weird here, feels like wasting a cycle of loop
-        }
-    });
-    // ans.clone().into_iter().for_each(|(ref_id, samples)| {
-    //     println!("ref_id: {}", ref_id);
-    //     println!("samples: {}", samples.len());
-    // });
-    print!("ans: {:#?}", ans);
-    ans
 }
 
 #[cfg(test)]
