+++ static_ri/Cargo.toml
@@ -12,4 +12,5 @@ dace_tests = { path = "../dace_tests" }
 tracing = "0.1.37"
 fxhash = "0.2.1"
 tracing-subscriber = "0.3.17"
+rand = "0.8.5"
